WpContext // fix isLoginRequest() to use correct new WP 6.1 Core function "is_login".
Since WordPress 6.1 there is a new Core function is_login() which will determine if the current request is going to the WordPress login page. Originally we checked for "is_login_screen()" but WordPress decided to name that function differently.

See: https://make.wordpress.org/core/2022/09/11/new-is_login-function-for-determining-if-a-page-is-the-login-screen/
Core Trac: https://core.trac.wordpress.org/ticket/19898